[FEATURE] split ConsensusXML and ConsensusXMLHandler
