+++ src/tests/class_tests/openms/source/ConsensusXMLFile_test.cpp
@@ -40,6 +40,7 @@
 ///////////////////////////
 
 #include <OpenMS/KERNEL/StandardTypes.h>
+#include <OpenMS/KERNEL/ConsensusMap.h>
 #include <OpenMS/FORMAT/FileHandler.h>
 #include <OpenMS/KERNEL/MSExperiment.h>
